Some changes to client side UI
Updated EmojinValinta -view according to new UI design. The counter for
the answers is not implemented yet. Currently it is not possible to test
with more than two emojis, so I wasn't able to test how the UI behaves
then.

Added fonts that are used in the new UI design.

Client results -view was updated. It now displays the results in graph
but the graph is upside down. Now the results view also triggers the client
to listen for the activity voting.