Add testMode to support development
Added testMode to support development with multiple Android emulators.
Normally client uses UDP communication to join the survey. This creates
challenge when developer wants to test the survey with separate host and
client emulators at the same time, because UDP communication between
Android emulators is not so straight forward. This is why testMode skips
the UDP discovery and tries to join the survey in known IP and port.
TestMode is boolean variable that can be changed in main.cs.

Note that the TCP connection requires port forward for the host
emulator: port 8000 -> 8001.