Fixes to activity voting and other improvements
These changes are mainly related to the activity vote process but this
contains also other small improvements. Activity voting is not fully
finished yet as the results are not correctly sent and the result
calculation needs to be implemented.

This commit add icon to the activity vote. Previously it was just
string, but now it is changed to Activity object. This enables adding an
image for the activity that is displayed to the client in voting phase.

Activity vote had previously two phases but the second phase is now
removed. There can be still traces of that in the code but those will be
removed in the future.

There was some improvements for the survey host and client. Previously
the host did not provide any information about the length of the content
it was sending to client. This lead to misinterpretations in client
sometimes and for example the emojis were read as part of the intro.

The client emoji voting view also now displays the correct question and
the activity vote options produce sound when the button is pressed.

Finally there is also small fix to the custom activity, as previosuly
the custom activity had to be created first.

This commit is too long..