Main screen UI changes and improvements to activity voting
Main screen UI was updated to better match the design. Also activity
voting was improved. If client does not receive results, user is
directed back to main screen with error alert. Also when client(s) do
not answer to the survey, the results are still sent from the host and
client can redirect to the result screen.